+++ core/server/master/src/test/java/alluxio/master/file/meta/LockedInodePathTest.java
@@ -15,21 +15,30 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.mock;
 
 import alluxio.AlluxioURI;
 import alluxio.TestLoggerRule;
+import alluxio.conf.Configuration;
+import alluxio.conf.PropertyKey;
 import alluxio.exception.InvalidPathException;
+import alluxio.exception.status.UnavailableException;
 import alluxio.master.file.meta.InodeTree.LockPattern;
+import alluxio.master.journal.JournalContext;
+import alluxio.master.journal.NoopJournalContext;
 
 import io.netty.util.ResourceLeakDetector;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
+import org.mockito.Mockito;
 
 import java.util.Arrays;
 import java.util.concurrent.ThreadLocalRandom;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Unit tests for {@link LockedInodePath}.
@@ -52,7 +61,10 @@ public class LockedInodePathTest extends BaseInodeLockingTest {
   public void pathExistsReadLock() throws Exception {
     AlluxioURI uri = new AlluxioURI("/a/b/c");
     mPath =
-        new LockedInodePath(uri, mInodeStore, mInodeLockManager, mRootDir, LockPattern.READ, false);
+        new LockedInodePath(
+            uri, mInodeStore, mInodeLockManager, mRootDir,
+            LockPattern.READ, false, NoopJournalContext.INSTANCE
+        );
     assertEquals(uri, mPath.getUri());
     assertEquals(4, mPath.size());
 
@@ -611,9 +623,37 @@ public class LockedInodePathTest extends BaseInodeLockingTest {
         + "resource is garbage-collected"));
   }
 
+  @Test
+  public void testFlushJournal() throws UnavailableException, InvalidPathException {
+    AtomicInteger journalFlushCount = new AtomicInteger();
+    JournalContext journalContext = mock(JournalContext.class);
+    Mockito.doAnswer(
+        (mock) -> {
+          journalFlushCount.getAndIncrement();
+          return null;
+        }
+    ).when(journalContext).flush();
+    Configuration.set(
+        PropertyKey.MASTER_FILE_SYSTEM_MERGE_INODE_JOURNALS,
+        true);
+
+    try (LockedInodePath path = create("/a/missing", LockPattern.WRITE_EDGE, journalContext)) {
+      InodeFile inodeB = inodeFile(10, mDirA.getId(), "missing");
+      path.addNextInode(inodeB);
+      path.downgradeToRead();
+    }
+    Assert.assertEquals(3, journalFlushCount.get());
+  }
+
   private LockedInodePath create(String path, LockPattern lockPattern) throws InvalidPathException {
+    return create(path, lockPattern, NoopJournalContext.INSTANCE);
+  }
+
+  private LockedInodePath create(
+      String path, LockPattern lockPattern, JournalContext journalContext
+  ) throws InvalidPathException {
     LockedInodePath lockedPath = new LockedInodePath(new AlluxioURI(path), mInodeStore,
-        mInodeLockManager, mRootDir, lockPattern, false);
+        mInodeLockManager, mRootDir, lockPattern, false, journalContext);
     lockedPath.traverse();
     return lockedPath;
   }