+++ core/server/worker/src/main/java/alluxio/worker/block/MonoBlockStore.java
@@ -12,13 +12,15 @@
 package alluxio.worker.block;
 
 import static alluxio.worker.block.BlockMetadataManager.WORKER_STORAGE_TIER_ASSOC;
+import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 
-import alluxio.Constants;
-import alluxio.exception.AlluxioRuntimeException;
-import alluxio.exception.BlockDoesNotExistRuntimeException;
-import alluxio.exception.ExceptionMessage;
-import alluxio.exception.WorkerOutOfSpaceException;
+import alluxio.conf.Configuration;
+import alluxio.conf.PropertyKey;
+import alluxio.exception.runtime.AlluxioRuntimeException;
+import alluxio.exception.runtime.BlockDoesNotExistRuntimeException;
+import alluxio.exception.runtime.DeadlineExceededRuntimeException;
+import alluxio.exception.status.AlluxioStatusException;
 import alluxio.exception.status.NotFoundException;
 import alluxio.exception.status.UnavailableException;
 import alluxio.grpc.Block;
@@ -28,6 +30,7 @@ import alluxio.retry.ExponentialBackoffRetry;
 import alluxio.retry.RetryUtils;
 import alluxio.underfs.UfsManager;
 import alluxio.util.IdUtils;
+import alluxio.util.ThreadFactoryUtils;
 import alluxio.worker.block.io.BlockReader;
 import alluxio.worker.block.io.BlockWriter;
 import alluxio.worker.block.io.DelegatingBlockReader;
@@ -43,15 +46,16 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.OptionalInt;
 import java.util.OptionalLong;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Collectors;
 
 /**
  * A implementation of BlockStore.
@@ -59,11 +63,14 @@ import java.util.stream.Collectors;
  */
 public class MonoBlockStore implements BlockStore {
   private static final Logger LOG = LoggerFactory.getLogger(MonoBlockStore.class);
-
+  private static final long LOAD_TIMEOUT =
+      Configuration.getMs(PropertyKey.USER_NETWORK_RPC_KEEPALIVE_TIMEOUT);
   private final LocalBlockStore mLocalBlockStore;
   private final UnderFileSystemBlockStore mUnderFileSystemBlockStore;
   private final BlockMasterClientPool mBlockMasterClientPool;
   private final AtomicReference<Long> mWorkerId;
+  private final ScheduledExecutorService mDelayer =
+      new ScheduledThreadPoolExecutor(1, ThreadFactoryUtils.build("LoadTimeOut", true));
 
   /**
    * Constructor of MonoBlockStore.
@@ -85,7 +92,7 @@ public class MonoBlockStore implements BlockStore {
   }
 
   @Override
-  public void abortBlock(long sessionId, long blockId) throws IOException {
+  public void abortBlock(long sessionId, long blockId) {
     mLocalBlockStore.abortBlock(sessionId, blockId);
   }
 
@@ -100,8 +107,7 @@ public class MonoBlockStore implements BlockStore {
   }
 
   @Override
-  public void commitBlock(long sessionId, long blockId, boolean pinOnCreate)
-      throws IOException {
+  public void commitBlock(long sessionId, long blockId, boolean pinOnCreate) {
     OptionalLong lockId = OptionalLong.of(
         mLocalBlockStore.commitBlockLocked(sessionId, blockId, pinOnCreate));
 
@@ -114,8 +120,8 @@ public class MonoBlockStore implements BlockStore {
       blockMasterClient.commitBlock(mWorkerId.get(),
           mLocalBlockStore.getBlockStoreMeta().getUsedBytesOnTiers().get(loc.tierAlias()),
           loc.tierAlias(), loc.mediumType(), blockId, meta.getBlockSize());
-    } catch (Exception e) {
-      throw new IOException(ExceptionMessage.FAILED_COMMIT_BLOCK_TO_MASTER.getMessage(blockId), e);
+    } catch (AlluxioStatusException e) {
+      throw AlluxioRuntimeException.from(e);
     } finally {
       mBlockMasterClientPool.release(blockMasterClient);
       if (lockId.isPresent()) {
@@ -127,8 +133,7 @@ public class MonoBlockStore implements BlockStore {
 
   @Override
   public String createBlock(long sessionId, long blockId, int tier,
-      CreateBlockOptions createBlockOptions)
-      throws WorkerOutOfSpaceException, IOException {
+      CreateBlockOptions createBlockOptions) {
     BlockStoreLocation loc;
     String tierAlias = WORKER_STORAGE_TIER_ASSOC.getAlias(tier);
     if (Strings.isNullOrEmpty(createBlockOptions.getMedium())) {
@@ -178,7 +183,7 @@ public class MonoBlockStore implements BlockStore {
       } catch (Exception ee) {
         LOG.warn("Failed to close UFS block", ee);
       }
-      String errorMessage = String.format("Failed to read from UFS, sessionId=%d, "
+      String errorMessage = format("Failed to read from UFS, sessionId=%d, "
               + "blockId=%d, offset=%d, positionShort=%s, options=%s: %s",
           sessionId, blockId, offset, positionShort, options, e);
       if (e instanceof FileNotFoundException) {
@@ -245,7 +250,7 @@ public class MonoBlockStore implements BlockStore {
 
   @Override
   public void moveBlock(long sessionId, long blockId, AllocateOptions moveOptions)
-      throws WorkerOutOfSpaceException, IOException {
+      throws IOException {
     mLocalBlockStore.moveBlock(sessionId, blockId, moveOptions);
   }
 
@@ -280,70 +285,82 @@ public class MonoBlockStore implements BlockStore {
   }
 
   @Override
-  public void requestSpace(long sessionId, long blockId, long additionalBytes)
-      throws WorkerOutOfSpaceException, IOException {
+  public void requestSpace(long sessionId, long blockId, long additionalBytes) {
     mLocalBlockStore.requestSpace(sessionId, blockId, additionalBytes);
   }
 
   @Override
-  public List<BlockStatus> load(List<Block> blocks, String tag, OptionalInt bandwidth) {
-    ArrayList<CompletableFuture<BlockStatus>> futures = new ArrayList<>();
+  public CompletableFuture<List<BlockStatus>> load(List<Block> blocks, String tag,
+      OptionalLong bandwidth) {
+    ArrayList<CompletableFuture<Void>> futures = new ArrayList<>();
+    List<BlockStatus> errors = Collections.synchronizedList(new ArrayList<>());
+    long sessionId = IdUtils.createSessionId();
     for (Block block : blocks) {
-      CompletableFuture<Void> future = CompletableFuture.runAsync(() -> {
-        try {
-          loadInternal(block.getBlockId(), block.getBlockSize(), block.getMountId(),
-              block.getUfsPath(), tag, bandwidth, block.getOffsetInFile());
-        } catch (Exception e) {
-          throw AlluxioRuntimeException.from(e);
+      long blockId = block.getBlockId();
+      long blockSize = block.getLength();
+      BlockWriter blockWriter;
+      UfsIOManager manager;
+      BlockStoreLocation loc =
+          BlockStoreLocation.anyDirInTier(WORKER_STORAGE_TIER_ASSOC.getAlias(0));
+      try {
+        manager = mUnderFileSystemBlockStore.getOrAddUfsIOManager(block.getMountId());
+        if (bandwidth.isPresent()) {
+          manager.setQuota(tag, bandwidth.getAsLong());
         }
-      }, GrpcExecutors.BLOCK_READER_EXECUTOR);
-      CompletableFuture<BlockStatus> exceptionally = future.handle((res, throwable) -> {
-        if (throwable != null) {
-          return BlockStatus.newBuilder().setBlock(block)
-              .setCode(AlluxioRuntimeException.from(throwable).getStatus().getCode().value())
-              .setMessage(throwable.getMessage()).setRetryable(false).build();
-        } else {
-          return null;
-        }
-      });
-      futures.add(exceptionally);
+        mLocalBlockStore.createBlock(sessionId, blockId, AllocateOptions.forCreate(blockSize, loc));
+        blockWriter = mLocalBlockStore.createBlockWriter(sessionId, blockId);
+      } catch (Exception e) {
+        handleException(e, block, errors, sessionId);
+        continue;
+      }
+      CompletableFuture<Void> future = RetryUtils.retryCallable("read from ufs",
+              () -> manager.read(blockId, block.getOffsetInFile(), blockSize, block.getUfsPath(),
+                  false, tag), new ExponentialBackoffRetry(1000, 5000, 5))
+          // use orTimeout in java 11
+          .applyToEither(timeoutAfter(LOAD_TIMEOUT, TimeUnit.MILLISECONDS), d -> d)
+          .thenAcceptAsync(d -> blockWriter.append(ByteBuffer.wrap(d)),
+              GrpcExecutors.BLOCK_WRITER_EXECUTOR)
+          .thenRun(() -> {
+            try {
+              blockWriter.close();
+            } catch (IOException e) {
+              throw AlluxioRuntimeException.from(e);
+            }
+          })
+          .thenRun(() -> commitBlock(sessionId, blockId, false))
+          .exceptionally(t -> {
+            handleException(t.getCause(), block, errors, sessionId);
+            return null;
+          });
+      futures.add(future);
     }
-    return futures.stream().map(CompletableFuture::join).filter(Objects::nonNull)
-        .collect(Collectors.toList());
+    return CompletableFuture.allOf(futures.toArray(new CompletableFuture[0]))
+        .thenApply(x -> errors);
   }
 
-  private void loadInternal(long blockId, long blockSize, long mountId, String ufsPath, String tag,
-      OptionalInt bandwidth, long offsetInUfs) throws WorkerOutOfSpaceException, IOException {
-    UfsIOManager manager = mUnderFileSystemBlockStore.getOrAddUfsIOManager(mountId);
-    if (bandwidth.isPresent()) {
-      manager.setQuotaInMB(tag, bandwidth.getAsInt());
+  private void handleException(Throwable e, Block block, List<BlockStatus> errors, long sessionId) {
+    AlluxioRuntimeException exception = AlluxioRuntimeException.from(e);
+    BlockStatus.Builder builder = BlockStatus.newBuilder().setBlock(block)
+        .setCode(exception.getStatus().getCode().value()).setRetryable(exception.isRetryable());
+    if (exception.getMessage() != null) {
+      builder.setMessage(exception.getMessage());
     }
-    long sessionId = IdUtils.createSessionId();
-    BlockStoreLocation loc = BlockStoreLocation.anyDirInTier(WORKER_STORAGE_TIER_ASSOC.getAlias(0));
-    mLocalBlockStore.createBlock(sessionId, blockId, AllocateOptions.forCreate(blockSize, loc));
-    requestSpace(sessionId, blockId, blockSize);
-    try (BlockWriter blockWriter = mLocalBlockStore.createBlockWriter(sessionId, blockId)) {
-      long offset = 0;
-      while (offset < blockSize) {
-        long bufferSize = Math.min(8L * Constants.MB, blockSize - offset);
-        long currentOffset = offset;
-        CompletableFuture<byte[]> data = RetryUtils.retryCallable("read from ufs",
-            () -> manager.read(blockId, offsetInUfs + currentOffset, bufferSize, ufsPath,
-                false, tag),
-            new ExponentialBackoffRetry(1000, 5000, 5));
-        offset += bufferSize;
-        ByteBuffer buffer = ByteBuffer.wrap(data.join());
-        blockWriter.append(buffer);
-      }
-    } catch (Exception e) {
+    errors.add(builder.build());
+    if (hasTempBlockMeta(block.getBlockId())) {
       try {
-        abortBlock(sessionId, blockId);
-      } catch (IOException ee) {
-        LOG.error("Failed to abort block after failing block write:", ee);
+        abortBlock(sessionId, block.getBlockId());
+      } catch (Exception ee) {
+        LOG.warn(format("fail to abort temp block %s after failing to load block",
+            block.getBlockId()), ee);
       }
-      throw e;
     }
-    commitBlock(sessionId, blockId, false);
+  }
+
+  private <T> CompletableFuture<T> timeoutAfter(long timeout, TimeUnit unit) {
+    CompletableFuture<T> result = new CompletableFuture<>();
+    mDelayer.schedule(() -> result.completeExceptionally(new DeadlineExceededRuntimeException(
+        format("time out after waiting for %s %s", timeout, unit))), timeout, unit);
+    return result;
   }
 
   @Override