+++ core/server/worker/src/test/java/alluxio/worker/block/DefaultBlockWorkerTest.java
@@ -36,10 +36,12 @@ import alluxio.Constants;
 import alluxio.Sessions;
 import alluxio.conf.Configuration;
 import alluxio.conf.PropertyKey;
-import alluxio.exception.BlockDoesNotExistRuntimeException;
-import alluxio.exception.WorkerOutOfSpaceException;
+import alluxio.exception.runtime.AlluxioRuntimeException;
+import alluxio.exception.runtime.BlockDoesNotExistRuntimeException;
+import alluxio.exception.runtime.ResourceExhaustedRuntimeException;
 import alluxio.exception.status.DeadlineExceededException;
 import alluxio.exception.status.NotFoundException;
+import alluxio.exception.status.UnavailableException;
 import alluxio.grpc.Block;
 import alluxio.grpc.BlockStatus;
 import alluxio.grpc.CacheRequest;
@@ -78,9 +80,10 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 import java.util.Random;
 import java.util.Set;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.atomic.AtomicReference;
 
 /**
@@ -246,7 +249,7 @@ public class DefaultBlockWorkerTest {
     long sessionId = mRandom.nextLong();
 
     // simulate server failure to commit block
-    doThrow(IOException.class)
+    doThrow(new UnavailableException("test"))
         .when(mBlockMasterClient)
         .commitBlock(
             anyLong(),
@@ -261,7 +264,8 @@ public class DefaultBlockWorkerTest {
         blockId,
         0,
         new CreateBlockOptions(null, Constants.MEDIUM_MEM, 1));
-    assertThrows(IOException.class, () -> mBlockWorker.commitBlock(sessionId, blockId, false));
+    assertThrows(AlluxioRuntimeException.class,
+        () -> mBlockWorker.commitBlock(sessionId, blockId, false));
   }
 
   @Test
@@ -280,11 +284,12 @@ public class DefaultBlockWorkerTest {
     long ufsBlockLength = 1024;
 
     // simulate server failure to commit ufs block
-    doThrow(IOException.class)
+    doThrow(new UnavailableException("test"))
         .when(mBlockMasterClient)
         .commitBlockInUfs(anyLong(), anyLong());
 
-    assertThrows(IOException.class, () -> mBlockWorker.commitBlockInUfs(blockId, ufsBlockLength));
+    assertThrows(AlluxioRuntimeException.class,
+        () -> mBlockWorker.commitBlockInUfs(blockId, ufsBlockLength));
   }
 
   @Test
@@ -298,9 +303,9 @@ public class DefaultBlockWorkerTest {
   }
 
   @Test
-  public void createBlockOutOfSpace() throws Exception {
+  public void createBlockOutOfSpace() {
     // simulates worker out of space
-    doThrow(WorkerOutOfSpaceException.class)
+    doThrow(ResourceExhaustedRuntimeException.class)
         .when(mBlockStore)
         .createBlock(anyLong(), anyLong(), anyInt(), any(CreateBlockOptions.class));
 
@@ -308,7 +313,7 @@ public class DefaultBlockWorkerTest {
     long blockId = mRandom.nextLong();
 
     assertThrows(
-        WorkerOutOfSpaceException.class,
+        ResourceExhaustedRuntimeException.class,
         () -> mBlockWorker.createBlock(
             sessionId,
             blockId,
@@ -404,7 +409,7 @@ public class DefaultBlockWorkerTest {
   }
 
   @Test
-  public void requestSpace() throws Exception {
+  public void requestSpace() {
     long blockId = mRandom.nextLong();
     long sessionId = mRandom.nextLong();
     long initialBytes = 512;
@@ -426,12 +431,12 @@ public class DefaultBlockWorkerTest {
   }
 
   @Test
-  public void requestSpaceNoSpace() throws Exception {
+  public void requestSpaceNoSpace() {
     long blockId = mRandom.nextLong();
     long sessionId = mRandom.nextLong();
     long additionalBytes = 2L * Constants.GB + 1;
     mBlockWorker.createBlock(sessionId, blockId, 1, new CreateBlockOptions(null, "", 1));
-    assertThrows(WorkerOutOfSpaceException.class,
+    assertThrows(ResourceExhaustedRuntimeException.class,
         () -> mBlockWorker.requestSpace(sessionId, blockId, additionalBytes)
     );
   }
@@ -470,16 +475,16 @@ public class DefaultBlockWorkerTest {
   }
 
   @Test
-  public void loadMultipleFromUfs() throws IOException {
+  public void loadMultipleFromUfs() throws IOException, ExecutionException, InterruptedException {
     Block block =
-        Block.newBuilder().setBlockId(0).setBlockSize(BLOCK_SIZE)
+        Block.newBuilder().setBlockId(0).setLength(BLOCK_SIZE)
             .setMountId(UFS_LOAD_MOUNT_ID).setOffsetInFile(0).setUfsPath(mTestLoadFilePath).build();
-    Block block2 = Block.newBuilder().setBlockId(1).setBlockSize(BLOCK_SIZE / 2)
+    Block block2 = Block.newBuilder().setBlockId(1).setLength(BLOCK_SIZE / 2)
         .setMountId(UFS_LOAD_MOUNT_ID).setOffsetInFile(BLOCK_SIZE)
         .setUfsPath(mTestLoadFilePath).build();
 
     List<BlockStatus> res =
-        mBlockWorker.load(Arrays.asList(block, block2), "test", OptionalInt.empty());
+        mBlockWorker.load(Arrays.asList(block, block2), "test", OptionalLong.empty()).get();
     assertEquals(res.size(), 0);
     assertTrue(mBlockStore.hasBlockMeta(0));
     assertTrue(mBlockStore.hasBlockMeta(1));
@@ -495,15 +500,15 @@ public class DefaultBlockWorkerTest {
   }
 
   @Test
-  public void loadDuplicateBlock() {
+  public void loadDuplicateBlock() throws ExecutionException, InterruptedException {
     int blockId = 0;
-    Block blocks = Block.newBuilder().setBlockId(blockId).setBlockSize(BLOCK_SIZE)
+    Block blocks = Block.newBuilder().setBlockId(blockId).setLength(BLOCK_SIZE)
         .setMountId(UFS_LOAD_MOUNT_ID).setOffsetInFile(0).setUfsPath(mTestLoadFilePath).build();
     List<BlockStatus> res =
-        mBlockWorker.load(Collections.singletonList(blocks), "test", OptionalInt.empty());
+        mBlockWorker.load(Collections.singletonList(blocks), "test", OptionalLong.empty()).get();
     assertEquals(res.size(), 0);
     List<BlockStatus> failure =
-        mBlockWorker.load(Collections.singletonList(blocks), "test", OptionalInt.empty());
+        mBlockWorker.load(Collections.singletonList(blocks), "test", OptionalLong.empty()).get();
     assertEquals(failure.size(), 1);
   }
 
@@ -668,8 +673,12 @@ public class DefaultBlockWorkerTest {
 
     long sessionId = mRandom.nextLong();
     // check that we can read the block locally
+    // note: this time we use an OpenUfsOption without ufsPath and blockInUfsTier so
+    // that the worker can't fall back to ufs read.
+    Protocol.OpenUfsBlockOptions noFallbackOptions = Protocol.OpenUfsBlockOptions.newBuilder()
+        .setBlockInUfsTier(false).build();
     try (BlockReader reader = mBlockWorker.createBlockReader(
-            sessionId, blockId, 0, false, options)) {
+            sessionId, blockId, 0, false, noFallbackOptions)) {
       ByteBuffer buf = reader.read(0, ufsBlockSize);
       // alert: LocalFileBlockReader uses a MappedByteBuffer, which does not
       // support the array operation. So we need to compare ByteBuffer manually