+++ integration/fuse/src/main/java/alluxio/fuse/file/FuseFileStream.java
@@ -105,6 +105,15 @@ public interface FuseFileStream extends AutoCloseable {
     public FuseFileStream create(
         AlluxioURI uri, int flags, long mode) {
       Optional<URIStatus> status = AlluxioFuseUtils.getPathStatus(mFileSystem, uri);
+      if (status.isPresent() && !status.get().isCompleted()) {
+        // Fuse.release() is async
+        // added for write-then-read and write-then-overwrite workloads
+        status = AlluxioFuseUtils.waitForFileCompleted(mFileSystem, uri);
+        if (!status.isPresent()) {
+          throw new UnsupportedOperationException(String.format(
+              "Failed to create fuse file stream for %s: file is being written", uri));
+        }
+      }
       switch (OpenFlags.valueOf(flags & O_ACCMODE.intValue())) {
         case O_RDONLY:
           return FuseFileInStream.create(mFileSystem, uri, status);