+++ stress/shell/src/main/java/alluxio/stress/cli/UfsIOBench.java
@@ -11,13 +11,13 @@
 
 package alluxio.stress.cli;
 
-import alluxio.cli.ValidationUtils;
 import alluxio.conf.Configuration;
 import alluxio.stress.worker.IOTaskResult;
 import alluxio.stress.worker.UfsIOParameters;
 import alluxio.underfs.UnderFileSystem;
 import alluxio.underfs.UnderFileSystemConfiguration;
 import alluxio.util.CommonUtils;
+import alluxio.util.ExceptionUtils;
 import alluxio.util.FormatUtils;
 import alluxio.util.executor.ExecutorServiceFactories;
 import alluxio.util.io.PathUtils;
@@ -97,7 +97,7 @@ public class UfsIOBench extends Benchmark<IOTaskResult> {
         result = new IOTaskResult();
         result.setParameters(mParameters);
         result.setBaseParameters(mBaseParameters);
-        result.addError(ValidationUtils.getErrorInfo(e));
+        result.addError(ExceptionUtils.asPlainText(e));
       }
       return result;
     } finally {
@@ -178,7 +178,7 @@ public class UfsIOBench extends Benchmark<IOTaskResult> {
       IOTaskResult result = new IOTaskResult();
       result.setParameters(mParameters);
       result.setBaseParameters(mBaseParameters);
-      result.addError(ValidationUtils.getErrorInfo(e));
+      result.addError(ExceptionUtils.asPlainText(e));
       return result;
     }
 
@@ -212,7 +212,7 @@ public class UfsIOBench extends Benchmark<IOTaskResult> {
           LOG.debug("Read task finished {}", p);
         } catch (Exception e) {
           LOG.error("Failed to read {}", filePath, e);
-          result.addError(ValidationUtils.getErrorInfo(e));
+          result.addError(ExceptionUtils.asPlainText(e));
         } finally {
           if (inStream != null) {
             try {
@@ -264,7 +264,7 @@ public class UfsIOBench extends Benchmark<IOTaskResult> {
       IOTaskResult result = new IOTaskResult();
       result.setParameters(mParameters);
       result.setBaseParameters(mBaseParameters);
-      result.addError(ValidationUtils.getErrorInfo(e));
+      result.addError(ExceptionUtils.asPlainText(e));
       return result;
     }