+++ tests/src/test/java/alluxio/client/rest/S3ClientRestApiTest.java
@@ -57,13 +57,13 @@ import alluxio.util.CommonUtils;
 import alluxio.wire.FileInfo;
 
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
+import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.io.BaseEncoding;
 import org.apache.commons.codec.binary.Hex;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
-import org.apache.hadoop.hbase.util.Strings;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -110,7 +110,7 @@ public final class S3ClientRestApiTest extends RestApiTest {
       .setProperty(PropertyKey.SECURITY_AUTHORIZATION_PERMISSION_ENABLED, false)
       .setProperty(PropertyKey.SECURITY_AUTHENTICATION_TYPE, AuthType.NOSASL)
       .setProperty(PropertyKey.USER_FILE_BUFFER_BYTES, "1KB")
-      .setProperty(PropertyKey.PROXY_S3_MULTIPART_UPLOAD_MIN_PART_SIZE, "0")
+      .setProperty(PropertyKey.PROXY_S3_COMPLETE_MULTIPART_UPLOAD_MIN_PART_SIZE, "0")
       .setProperty(PropertyKey.PROXY_S3_TAGGING_RESTRICTIONS_ENABLED, true) // default
       .setProperty(PropertyKey.PROXY_S3_BUCKET_NAMING_RESTRICTIONS_ENABLED, false) // default
       .setProperty(PropertyKey.PROXY_S3_MULTIPART_UPLOAD_CLEANER_ENABLED, false)
@@ -250,7 +250,7 @@ public final class S3ClientRestApiTest extends RestApiTest {
         TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
         .runAndCheckResult(expected);
 
-    //parameters with list-type=2 start-after="folder0/file0"
+    //parameters with list-type=2 start-after="file0"
     expected = new ListBucketResult("bucket", statuses,
         ListBucketOptions.defaults().setListType(2).setStartAfter("file0"));
     assertEquals(5, expected.getContents().size());
@@ -268,6 +268,25 @@ public final class S3ClientRestApiTest extends RestApiTest {
         "bucket", parameters, HttpMethod.GET,
         TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
         .runAndCheckResult(expected);
+  }
+
+  @Test
+  public void listNonExistentBucket() throws Exception {
+    String bucketName = "bucket";
+    //empty parameters
+    List<URIStatus> statuses = mFileSystem.listStatus(new AlluxioURI("/"),
+        ListStatusPOptions.newBuilder().setRecursive(true).build());
+
+    // Verify 404 HTTP status & NoSuchBucket S3 error code
+    HttpURLConnection connection = new TestCase(mHostname, mPort, mBaseUri,
+        bucketName, NO_PARAMS, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .execute();
+    Assert.assertEquals(404, connection.getResponseCode());
+    S3Error response =
+        new XmlMapper().readerFor(S3Error.class).readValue(connection.getErrorStream());
+    Assert.assertEquals(bucketName, response.getResource());
+    Assert.assertEquals(S3ErrorCode.Name.NO_SUCH_BUCKET, response.getCode());
   }
 
   @Test
@@ -610,6 +629,151 @@ public final class S3ClientRestApiTest extends RestApiTest {
     assertNull(expected.getCommonPrefixes());
 
     parameters.put("list-type", "2");
+    new TestCase(mHostname, mPort, mBaseUri,
+        "bucket", parameters, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .runAndCheckResult(expected);
+  }
+
+  @Test
+  public void listBucketPrefixZeroMatches() throws Exception {
+    // Test ListObjects with an empty bucket
+    mFileSystem.createDirectory(new AlluxioURI("/bucket"));
+    List<URIStatus> statuses = mFileSystem.listStatus(new AlluxioURI("/bucket"),
+        ListStatusPOptions.newBuilder().setRecursive(true).build());
+
+    //parameters with no prefix
+    ListBucketResult expected = new ListBucketResult("bucket", statuses,
+        ListBucketOptions.defaults());
+    assertEquals(0, expected.getContents().size());
+
+    final Map<String, String> parameters = new HashMap<>();
+    new TestCase(mHostname, mPort, mBaseUri,
+        "bucket", parameters, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .runAndCheckResult(expected);
+
+    //parameters with prefix=""
+    expected = new ListBucketResult("bucket", statuses,
+        ListBucketOptions.defaults().setPrefix(""));
+    assertEquals(0, expected.getContents().size());
+
+    parameters.put("prefix", "");
+    new TestCase(mHostname, mPort, mBaseUri,
+        "bucket", parameters, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .runAndCheckResult(expected);
+
+    // Test ListObjects with objects in the bucket
+    mFileSystem.createFile(new AlluxioURI("/bucket/file0"));
+    mFileSystem.createDirectory(new AlluxioURI("/bucket/file_store"));
+    mFileSystem.createFile(new AlluxioURI("/bucket/file_store/file1"));
+    statuses = mFileSystem.listStatus(new AlluxioURI("/bucket"),
+        ListStatusPOptions.newBuilder().setRecursive(true).build());
+
+    //parameters with no prefix
+    expected = new ListBucketResult("bucket", statuses,
+        ListBucketOptions.defaults());
+    assertEquals(3, expected.getContents().size());
+    assertEquals("file0", expected.getContents().get(0).getKey());
+    assertEquals("file_store/", expected.getContents().get(1).getKey());
+    assertEquals("file_store/file1", expected.getContents().get(2).getKey());
+
+    parameters.remove("prefix");
+    new TestCase(mHostname, mPort, mBaseUri,
+        "bucket", parameters, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .runAndCheckResult(expected);
+
+    //parameters with prefix=""
+    expected = new ListBucketResult("bucket", statuses,
+        ListBucketOptions.defaults().setPrefix(""));
+    assertEquals(3, expected.getContents().size());
+    assertEquals("file0", expected.getContents().get(0).getKey());
+    assertEquals("file_store/", expected.getContents().get(1).getKey());
+    assertEquals("file_store/file1", expected.getContents().get(2).getKey());
+
+    parameters.put("prefix", "");
+    new TestCase(mHostname, mPort, mBaseUri,
+        "bucket", parameters, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .runAndCheckResult(expected);
+
+    //parameters with non-existent prefix="dne_folder/file"
+    try {
+      expected = new ListBucketResult("bucket", statuses,
+          ListBucketOptions.defaults().setPrefix("dne_folder/file"));
+    } catch (Exception e) {
+      // expected
+      // TODO(czhu): with the current implementation of prefixes w/o delimiters, there is
+      // never a FileDoesNotExistException because we just list the entire bucket recursively
+      statuses = new ArrayList<>();
+      return;
+    }
+    assertEquals(0, expected.getContents().size());
+
+    parameters.put("prefix", "dne_folder/file");
+    new TestCase(mHostname, mPort, mBaseUri,
+        "bucket", parameters, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .runAndCheckResult(expected);
+
+    // prefix test with delimiter="/"
+    statuses = mFileSystem.listStatus(new AlluxioURI("/bucket"),
+        ListStatusPOptions.newBuilder().setRecursive(false).build());
+
+    //parameters with prefix="file"
+    expected = new ListBucketResult("bucket", statuses,
+        ListBucketOptions.defaults().setPrefix("file").setDelimiter(AlluxioURI.SEPARATOR));
+    assertEquals(1, expected.getContents().size());
+    assertEquals("file0", expected.getContents().get(0).getKey());
+    assertEquals(1, expected.getCommonPrefixes().size());
+    assertEquals("file_store/", expected.getCommonPrefixes().get(0).getPrefix());
+
+    parameters.put("delimiter", AlluxioURI.SEPARATOR);
+    parameters.put("prefix", "file");
+    new TestCase(mHostname, mPort, mBaseUri,
+        "bucket", parameters, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .runAndCheckResult(expected);
+
+    //parameters with prefix="file_store/file1/"
+    try {
+      statuses = mFileSystem.listStatus(new AlluxioURI("/bucket/file_store/file1/"),
+          ListStatusPOptions.newBuilder().setRecursive(false).build());
+    } catch (Exception e) {
+      // expected
+      // TODO(czhu): with the current implementation of listStatus() the trailing '/' character
+      // doesn't cause a failure, despite `/bucket/file_store/file1` not being a directory
+      statuses = new ArrayList<>();
+      return;
+    }
+    expected = new ListBucketResult("bucket", statuses, ListBucketOptions.defaults()
+        .setPrefix("file_store/file1/").setDelimiter(AlluxioURI.SEPARATOR));
+    assertEquals(0, expected.getContents().size());
+    assertEquals(0, expected.getCommonPrefixes().size());
+
+    parameters.put("prefix", "file_store/file1/");
+    new TestCase(mHostname, mPort, mBaseUri,
+        "bucket", parameters, HttpMethod.GET,
+        TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
+        .runAndCheckResult(expected);
+
+    //parameters with prefix="file_store/file2"
+    try {
+      statuses = mFileSystem.listStatus(new AlluxioURI("/bucket/file_store/file2"),
+          ListStatusPOptions.newBuilder().setRecursive(false).build());
+    } catch (FileDoesNotExistException e) {
+      // expected
+      statuses = new ArrayList<>();
+      return;
+    }
+    expected = new ListBucketResult("bucket", statuses, ListBucketOptions.defaults()
+        .setPrefix("file_store/file2").setDelimiter(AlluxioURI.SEPARATOR));
+    assertEquals(0, expected.getContents().size());
+    assertEquals(0, expected.getCommonPrefixes().size());
+
+    parameters.put("prefix", "file_store/file2");
     new TestCase(mHostname, mPort, mBaseUri,
         "bucket", parameters, HttpMethod.GET,
         TestCaseOptions.defaults().setContentType(TestCaseOptions.XML_CONTENT_TYPE))
@@ -1548,8 +1712,8 @@ public final class S3ClientRestApiTest extends RestApiTest {
     String objectKey = bucketName + AlluxioURI.SEPARATOR + objectName;
     String objectData = CommonUtils.randomAlphaNumString(DATA_SIZE);
 
-    String longTagKey = Strings.repeat('a', 128 + 1);
-    String longTagValue = Strings.repeat('b', 256 + 1);
+    String longTagKey = Strings.repeat("a", 128 + 1);
+    String longTagValue = Strings.repeat("b", 256 + 1);
     createObjectRestCall(objectKey, NO_PARAMS,
         TestCaseOptions.defaults()
             .setBody(objectData.getBytes())