+++ underfs/hdfs/src/main/java/alluxio/underfs/hdfs/HdfsUnderFileInputStream.java
@@ -11,8 +11,6 @@
 
 package alluxio.underfs.hdfs;
 
-import static alluxio.underfs.hdfs.HdfsUnderFileSystem.tryHdfsCall;
-
 import alluxio.underfs.SeekableUnderFileInputStream;
 
 import org.apache.hadoop.fs.FSDataInputStream;
@@ -30,20 +28,12 @@ public class HdfsUnderFileInputStream extends SeekableUnderFileInputStream {
   }
 
   @Override
-  public void seek(long position) {
-    tryHdfsCall(() -> {
-      ((FSDataInputStream) in).seek(position);
-      return null;
-    });
+  public void seek(long position) throws IOException {
+    ((FSDataInputStream) in).seek(position);
   }
 
   @Override
   public long getPos() throws IOException {
     return ((FSDataInputStream) in).getPos();
-  }
-
-  @Override
-  public int read(byte[] b, int off, int len) {
-    return tryHdfsCall(() -> in.read(b, off, len));
   }
 }