+++ integration/kubernetes/helm-chart/alluxio/CHANGELOG.md
@@ -291,4 +291,13 @@
 - Support ConfigMap mounts and allow users to mount ConfigMap volumes similar to Secrets
 
 0.6.47
+
 - Support alluxio proxy
+
+0.6.48
+
+- Replace all default space size to "i" unit to avoid the unit discrepancy between Alluxio and Kubernetes
+
+0.6.49
+
+- Remove dummy example alluxio-fuse-client